test(http): fuzz the request parser and cover the module
Five units and a fuzz target. The parser's test is mostly the hostile cases,
because that is what the parser is for: traversal in both spellings, a request
carrying two framings, two Content-Lengths, an obsolete folded header, a length
that does not fit a u64, every bound, and a request line that is not one. The
auth test is the ways a token can fail to be ours, `alg: none` and a correctly
signed HS512 among them. The router test drives nya_http_router_dispatch with
no socket, which is why that function is public.

Three defects the sanitizers and AFL found on the way, all fixed here:

- _nya_type_try_parse_s128 wrote "-0" as `~0 + 1`, an unsigned overflow that
  wrapped back to the right answer by undefined means. Reachable from any JSON
  number, which is how the HTTP body parser found it. Now S128_MIN is built
  rather than negated into and everything else negates a value known to fit.
- The idle timeout subtracted a timestamp taken before the read from one taken
  during it, so an ordinary connection could look like it had been idle for
  five hundred years. The clock is now read after the work.
- A 401 added WWW-Authenticate before writing the problem body, which clears
  the headers. The header goes on after.

`./build run fuzz` could not run a target at all, which is why none of this was
caught earlier: the replay helpers are unused under AFL and failed -Werror, and
the sanitizer environment lacked the three settings afl-fuzz refuses to start
without. Both fixed, and .fuzz/ is ignored. 5.3 million executions over four
minutes produced the one saved crash, kept as a regression input.

gnyame starts the server when GNYAME_WEB_PORT names a port and mounts the
metrics and schema resources behind the logging layer. Unset, nothing binds,
allocates or registers, which is the caller the verification rule asks for and
the zero cost the design promised.

+++ src/nyangine/http/http_server.h
@@ -104,7 +104,7 @@
  * document repeatedly and never reads would be an unbounded allocation. This is where that stops:
  * four of the largest response, and then the connection goes.
  * */
-#define NYA_HTTP_MAX_PENDING_WRITE_BYTES (NYA_HTTP_MAX_RESPONSE_BYTES * 4)
+#define NYA_HTTP_MAX_PENDING_WRITE_BYTES ((u64)NYA_HTTP_MAX_RESPONSE_BYTES * 4ULL)
 
 /*
  * ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
